chore(suite-native): refactor ExchangeAccountCard to general TradingAccountCard
